refactor: rest Controller와 view Controller 분리
