+++ src/main/java/roomescape/controller/AdminController.java
@@ -11,27 +11,18 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
 import roomescape.service.ReservationService;
 
-@Controller
+@RestController
 @RequestMapping("/admin")
 public class AdminController {
     private final ReservationService reservationService;
 
     public AdminController(ReservationService reservationService) {
         this.reservationService = reservationService;
-    }
-
-    @GetMapping
-    public String mainPage() {
-        return "admin/index";
-    }
-
-    @GetMapping("/reservation")
-    public String reservationPage() {
-        return "admin/reservation-new";
     }
 
     @PostMapping("/reservations")
@@ -51,15 +42,5 @@ public class AdminController {
             @RequestParam LocalDate end
     ) {
         return reservationService.findByMemberAndThemeBetweenDates(memberId, themeId, start, end);
-    }
-
-    @GetMapping("/time")
-    public String reservationTimePage() {
-        return "admin/time";
-    }
-
-    @GetMapping("/theme")
-    public String themePage() {
-        return "admin/theme";
     }
 }
