+++ src/main/java/roomescape/controller/view/AdminPageController.java
@@ -1,0 +1,29 @@
+package roomescape.controller.view;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminPageController {
+    @GetMapping
+    public String mainPage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/reservation")
+    public String reservationPage() {
+        return "admin/reservation-new";
+    }
+
+    @GetMapping("/time")
+    public String reservationTimePage() {
+        return "admin/time";
+    }
+
+    @GetMapping("/theme")
+    public String themePage() {
+        return "admin/theme";
+    }
+}
